Replace movai icon in LoginModal, LoginPanel and Scene Loader components
