+++ src/Components/LoginForm/style.js
@@ -1,0 +1,17 @@
+const styles = theme => ({
+  root: {
+    padding: theme.spacing(4, 4),
+    borderRadius: 40
+  },
+  container: {
+    paddingTop: "5%"
+  },
+  logoImage: {
+    display: "block",
+    marginLeft: "auto",
+    marginRight: "auto",
+    width: "50%"
+  }
+});
+
+export { styles };
